Split tests to dedicated package fields
Pruned include whenever possible

TODO: tests for TrainableField with Dirichlet constraint.

Ref #7, #5.
