+++ src/neurom/fields/__init__.py
@@ -1,0 +1,3 @@
+from neurom.fields.field_base import FieldBase
+from neurom.fields.field import Field
+from neurom.fields.trainable_field import TrainableField
