+++ tests/unit/fields/test_field.py
@@ -3,17 +3,16 @@ import torch
 import torch.nn as nn
 
 # Import library modules
-from neurom.field import TrainableField
+from neurom.fields import Field
 from neurom.meshes import Topology
-from neurom.constraints import NoConstraint
 
 torch.set_default_dtype(torch.float32)
 
 
 @pytest.fixture
-def field_no_constraint():
+def field():
     """
-    Prepare what is needed to define a TrainableField:
+    Prepare what is needed to define a Field:
     * name = "test"
     * Simple topology: 3 elements with 4 nodes.
     * Values: [3., 7., 6., -5.]
@@ -23,20 +22,15 @@ def field_no_constraint():
     nodes = torch.arange(0, N)
     elements = torch.vstack([torch.arange(0, N - 1), torch.arange(1, N)]).T
     topology = Topology(nodes, elements)
-    init_values = torch.tensor([3.0, 7.0, 6.0, -5.0])
-    field = TrainableField(
-        name="test",
-        topology=topology,
-        init_values=init_values,
-        constraint=NoConstraint(),
-    )
+    values = torch.tensor([3.0, 7.0, 6.0, -5.0])
+    field = Field(name="test", topology=topology, values=values)
 
     return field
 
 
-class TestTrainableField:
+class TestField:
     """
-    Test TrainableField class
+    Test Field class
 
     Attributes:
         relative_tolerance (float): The relative tolerance used to compare floats.
@@ -44,33 +38,35 @@ class TestTrainableField:
 
     relative_tolerance: float = 1e-9
 
-    def test_construction(self, field_no_constraint):
+    def test_construction(self, field):
         """
-        Test construction of a TrainableField.
+        Test construction of a Field.
         """
-        assert field_no_constraint.name == "test"
-        assert isinstance(field_no_constraint.values_reduced, nn.Parameter)
+        assert field.name == "test"
+
+        assert "values" in field._buffers
+        assert isinstance(getattr(field, "values"), torch.Tensor)
 
         expected_values = torch.tensor([3.0, 7.0, 6.0, -5.0])
-        assert field_no_constraint.values_reduced.detach() == pytest.approx(
+        assert field.values == pytest.approx(
             expected_values, rel=self.relative_tolerance
         )
 
-    def test_full_values(self, field_no_constraint):
+    def test_full_values(self, field):
         """
-        Test method TrainableField.full_values()
+        Test method Field.full_values()
         """
         expected_values = torch.tensor([3.0, 7.0, 6.0, -5.0])
-        assert field_no_constraint.full_values().detach() == pytest.approx(
+        assert field.full_values() == pytest.approx(
             expected_values, rel=self.relative_tolerance
         )
 
-    def at_elements(self, field_no_constraint):
+    def at_elements(self, field):
         """
-        Test method TrainableField.at_elements()
+        Test method Field.at_elements()
         """
 
         expected_values = torch.tensor([[3.0, 7.0], [7.0, 6.0], [6.0, -5.0]])
-        assert field_no_constraint.at_elements() == pytest.approx(
+        assert field.at_elements() == pytest.approx(
             expected_values, rel=self.relative_tolerance
         )
